create redux auto complete form as reusable component and use for both individual and organized events
